Better Touchscreen Zooming (#2435)
* Touch panning & zooming for touch drawing

Changes TouchDrawingInputHandler to allow multitouch and use it to pan &
zoom; attempt to fix TouchInputHandler.cpp scrolling bug.

* Pressure inference implementation

Also, further tweaks to TouchInputHandler.cpp in attempt to fix scroll jumping
bug.

* Open and save settings.glade with newer Glade

Newer versions of glade change underscores to dashes when saving older
files. This can lead to merge conflicts.

* Settings for pressure, drawing, clearing PdfCache

PdfCache previously cleared on zoom. Further attempts to fix scrolling
bug in TouchInputHandler.cpp

* Fix transition between hand tool & draw (touch)

Switching from the hand tool to the draw tool required a two-finger
gesture to take effect -- Gtk was blocking one-finger gestures, using
them for momentum scrolling.

* Suggestions from code review.

Commenting, undoing unnecessary changes, ...

* Large document zooming: Stop view scroll up/down

Resolves an issue where zooming in/out would cause the view to scroll up
or down when editing a large document. This is fixed by making
ZoomControl.cpp take into account that vertical padding doesn't scale
with zoom.

Unfortunately, while this fixes the scrolling issue, a viewport that
moves up and down as the user zooms is still an issue. The difference
is that while previously, zooming caused the view to move significantly,
the view now remains roughly centered.

* Scroll at normal speed when zoom is off & multiple touches are used.

`TouchInputHandler::scrollMotion` states that it is only to be called if
there is a single input sequence. When zoom gestures are disabled,
however, it was called when two input sequences were active. This has
been fixed by checking whether there are multiple input sequences before
calling `scrollMotion`.

* Stop delaying start of stroke in TouchDrawingInputHandler

Seems to fix a bug where, at times, after pinch-zooming with
touch-drawing enabled, switching to mouse drawing without a button down
would create a preview of a stroke.

* Fix transition: two-finger scrolling to one-finger

Apply @eliasrg's b9c50aab62ce58d157b5960f8d9c23dd1f98e7c4. This issue
occured when panning with two fingers, then attempting to pan with the
second finger pressed after lifting up the first.

Fix touch drawing/panning transition on Wayland.

* Additional zoom settings.

Add settings for PDF recache zoom difference threshold. Also adds a
threshold for which zoom snaps to the initial zoom if the difference in
distances between the two touch points is less than some given
percentage of the original.

The goal of this feature is to make it easier to pan with two fingers in
large PDFs/other documents in which zooming is slow.

* Fix kinetic scrolling for devices with touch drawing disabled

Kinetic scrolling would only be enabled if the touchDrawingEnabled setting
was true.

* Change touch drawing tooltip with its dependent settings.

The touch drawing checkbox' tooltip (and sensitivity) changes based on the value of the
settings it depends on.

+++ src/gui/sidebar/previews/base/SidebarPreviewBase.cpp
@@ -128,6 +128,6 @@ auto SidebarPreviewBase::scrollToPreview(SidebarPreviewBase* sidebar) -> bool {
     return false;
 }
 
-void SidebarPreviewBase::pageDeleted(int page) {}
+void SidebarPreviewBase::pageDeleted(size_t page) {}
 
-void SidebarPreviewBase::pageInserted(int page) {}
+void SidebarPreviewBase::pageInserted(size_t page) {}
